Fix #1356: Add new entities to ES index from OpenMetadata server
