+++ catalog-rest-service/src/main/resources/elasticsearch/dashboard_index_mapping.json
@@ -1,7 +1,6 @@
 {
-  "mappings":{
     "properties": {
-      "dashboard_name": {
+      "name": {
         "type":"text"
       },
       "display_name": {
@@ -69,5 +68,4 @@
         "type": "long"
       }
     }
-  }
 }
