+++ catalog-rest-service/src/main/resources/elasticsearch/dbt_index_mapping.json
@@ -1,17 +1,7 @@
 {
-  "mappings":{
     "properties": {
-      "dbt_model_name": {
+      "name": {
         "type":"text"
-      },
-      "schema": {
-        "type":"text",
-        "analyzer": "simple",
-        "fields": {
-          "raw": {
-            "type": "keyword"
-          }
-        }
       },
       "display_name": {
         "type": "text"
@@ -63,5 +53,4 @@
         "type": "completion"
       }
     }
-  }
 }
