+++ catalog-rest-service/src/main/resources/elasticsearch/pipeline_index_mapping.json
@@ -1,7 +1,6 @@
 {
-  "mappings":{
     "properties": {
-      "pipeline_name": {
+      "name": {
         "type":"text"
       },
       "display_name": {
@@ -51,5 +50,4 @@
         "type": "completion"
       }
     }
-  }
 }
