+++ catalog-rest-service/src/main/resources/elasticsearch/table_index_mapping.json
@@ -1,17 +1,7 @@
 {
-  "mappings":{
     "properties": {
-      "table_name": {
+      "name": {
         "type":"text"
-      },
-      "schema": {
-        "type":"text",
-        "analyzer": "simple",
-        "fields": {
-          "raw": {
-            "type": "keyword"
-          }
-        }
       },
       "display_name": {
         "type": "text"
@@ -43,9 +33,6 @@
       },
       "tags": {
         "type": "keyword"
-      },
-      "badges": {
-        "type": "text"
       },
       "service": {
         "type": "keyword"
@@ -84,5 +71,4 @@
         "type": "long"
       }
     }
-  }
 }
